Introduce EventIdentifier and exposing on EventContext.
This identifier uniquely identifies every event, and is composed from
the microservice and tenant which committed the event and the event log
sequence number it got when it was committed to the event store.

These values are converted to bytes as they are transferred over gRPC
i.e. in little endian order, concatenated and converted to a string with
base64 encoding.